Create cloned_soldier_count.py
finding all the clone soldiers in a clone army parade formation. The clone soldiers are uniquely identified by a number within a specific range, and they can be represented by a square matrix. Our goal is to count the number of clones for each ID and return the statistical result in a dictionary format. The solution should be implemented in the count_clone_soldier(matrix: List[List[str]])